Extract Swift enum declarator partial
